requirements.txt
aiogram==3.22.0
aiohttp==3.9.0
python-dotenv==1.2.1
requests==2.32.5